clear up some codes to make project working seamlessly

+++ src/main/resources/db/migrations/V2__create_conference_schema.sql
@@ -1,5 +1,5 @@
 CREATE TABLE conferences (
-    conference_id INT PRIMARY KEY AUTO_INCREMENT,
+    conference_id BIGINT PRIMARY KEY AUTO_INCREMENT,
     name VARCHAR(255) NOT NULL,
     location VARCHAR(255) NOT NULL,
     start_date DATE NOT NULL,
